requirements.txt
requests
python-dotenv
openai
pytest
flask
gunicorn
google-cloud-secret-manager
pytz